requirements.txt
streamlit==1.33.0
google-cloud-aiplatform==1.47.0
google-cloud-documentai==2.27.0
google-cloud-dlp==3.15.0
google-cloud-speech==2.24.0
google-cloud-firestore==2.16.0
python-dotenv==1.0.1
pandas==2.2.1
jira==3.6.0
fpdf2==2.7.9
python-docx==1.1.2
openpyxl==3.1.2

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-html==4.1.1
pytest-xdist==3.5.0

# Code quality
flake8==6.1.0
mypy==1.7.1
black==23.11.0
isort==5.12.0

# Additional utilities
requests==2.31.0
responses==0.24.1